Changed how vehicle gravity functions
Previously vehicle gravity simulated real world gravity with a force being applied on the vertical axis. This force was positive or negative, depending on gravity being reversed or not.

Now vehicle gravity works by shooting a raycast out onto a surface and checking for the "Track" tag. If this tag is found, the force applied will keep the vehicle hovering above the track. The script also checks the rotation of the normal face hit by the raycast and adjusts the rotation of the vehicle accordingly (this should allow us to work with loops and corkscrews) but this functionality is implemented incredibly simplistically and will need a lot of testing and refinement.