Tracking switching now working on a button press
This is a prototype/proof of concept version of the "track switching" mechanic. By pressing "z" on the keyboard, the player can instantly switch which side of the track they are on and gravity continues to apply to the ship. This should work on corkscrews etc too, but this hasn't been tested yet.

Eventually this code will need to be refined so the switch is more fluid and not instantaneous.