"CameraFollow" script added, changes made to other scripts and scene
Added a camera follow script that smooths out the camera movement. This script functions based on some new gameobject that are child objects of the ship. These objects serve as a "hook" and "focal point" for the camera. Also made changes to the "VehicleGravity" script, mainly tidying code up and moving things into methods.