[0.0.4] Basic split screen added, changes to vehicle steering
Vehicles now steer based on a turning limit. This is a variable that increases as the ship velocity increases, meaning the ship has a faster turning circle if the ship is going slower.

Also added in basic split screen functionality now. Depending on how many players are selected, the cameras automatically set themselves up accordingly. If 2 players are set, the game works in "vertical split screen" mode, if 3 players are set an image is brought up to plug the black space left where player 4's camera would normally be.