[0.0.7] Multiple changes (read desc.)
Added ship stats as public variables so they can be changed in inspector (for testing). Added different ship types with different stats to test. Changed camera manager object to "interface manager" and changed camera manager script to interface manager so we can expand upon it in the future for all interface related code.