[0.0.10] Multiple changes (read desc.)
Camera transition added for gravity shift. When the player shifts gravity now the screen is flashed white briefly to mask the rotation + movement of the ship through the track. There is also a new stat added to the vehicle control script that dictates how often the ship can use it's gravity charge power.