[0.0.11] User interface coding started
Added in the base code and elements for the user interface. Currently working are gravity charges, lap timer, total timer and speed. Other elements are prepared so they can be easily linked when they have been added (for example, once we have laps added, lap counter is ready to be tied to that code).