[0.0.14] Respawn system added
Repsawn system has been coded in. The player now respawns at the last checkpoint (currently only forced respawning using the "back" button).

The burden of this on the player can be alleviated somewhat by ensuring there are plenty of checkpoints added into the map. When the ship respawns it matches the forward direction of the checkpoint and lifts off the track by 3 units, to ensure there are no clipping issues.