[0.0.15] Checkpoint based position system added
A series of tests that determine the player's position are now added to the game manager script. These checks first determine if the player is ahead on laps, then if the player is ahead on checkpoints before finally determining if they are closer to the next checkpoint or not. This then ranks the players by position and sends the information to the interface manager.