[0.0.16] Powerup items (framework) and speed boosters added
There are now functional items on the map that either supply the player with a weapon (a placeholder text only currently) or levels that weapon up (up to a max of level 3). The "Eraser" weapon is special in that it cannot be levelled up and it will only be given to the player in last place.

The item boxes for these powerups/levelups disappear for 2 seconds after being picked up. If a player picks a levelup box with no weapon, nothing happens. Also if a player picks up the weapon box while already in possession of a weapon (levelled up or not) it is replaced with a new, random weapon at level 1.

Track speed boosters have also been added to the game and function as intended. If the player passes over one they are entered into a state of "forced acceleration" and are propelled forward at higher than their usual max speed and acceleration. This element may need balancing if it proves to be too weak (or too strong) for certain ships as the current speed up is based on the current max ship speed/acceleration times by a factor.