[0.0.23] Multiple changes (read desc)
Checkpoints - Added checkpoints back in that were previously removed due to a bug. The code now spawns you at the last checkpoint passed and matches it's rotation on all axis before checking if the ship is "flipped" or not, and then rotating the ship correctly. Due to this change a new boolean has been added to the vehicle control script called "flipped" and switches on/off when a ship performs a gravity flip.

Catchup - If the player is in 2nd, 3rd or 4th they are now given some "catch up" speed to help keep them competitive. This will need balancing as the game is tested to ensure the effect is not too weak or too strong.

Track Layout - The track now has twice the amount of interactable objects on the screen (speedboosts and pick up boxes). The speedboosts have also had their colliders enlarged so they are easier to hit when moving over them.

Weapons - The missile weapon has been tweaked slighty to be heavier (helps sticking to the track). The collider for the missile has also been increased to make it easier to hit with the weapon. This may still need tweaking depending on how effective/ineffective the weapon is in testing.