Small fixes
Changed the way missile works as it was a child object of the ship that shot it, which messed with it's positioning when the ship moved. Now after the vehicle control script has been grabbed from the parent ship (for target aquisition when homing) it's parent is set back to null. This also fixes a problem with missiles being able to pick up weapon boxes.

Also fixed a small problem with the newly implemented weapon/level up boxes. Their pivot points were set incorrectly so they have been moved to a child of an empty game object so the bob/rotation script rotates them correctly.